Create Netflix Analytics
This was my first time using PostgreSQL dialect instead of standard. This project was me just working through a set of problems provided by Charlotte Chaze using a database and tables that she provided with data about Netflix shows/movies. I queried the data to pull all the movie entries, what the most recently added show/movie was, sorted the titles alphabetically, found the director for a specific movie, then found the entry for the movie with the oldest release date in the data set.